+++ src/Ormolu/Utils/Cabal.hs
@@ -141,8 +141,8 @@ parseCabalInfo cabalFileAsGiven sourceFileAsGiven = liftIO $ do
   CachedCabalFile {..} <- whenNothing (M.lookup cabalFile cabalCache) $ do
     cabalFileBs <- B.readFile cabalFile
     genericPackageDescription <-
-      whenNothing (parseGenericPackageDescriptionMaybe cabalFileBs) $
-        throwIO (OrmoluCabalFileParsingFailed cabalFile)
+      whenLeft (snd . runParseResult $ parseGenericPackageDescription cabalFileBs) $
+        throwIO . OrmoluCabalFileParsingFailed cabalFile . snd
     let extensionsAndDeps =
           getExtensionAndDepsMap cabalFile genericPackageDescription
         cachedCabalFile = CachedCabalFile {..}
@@ -163,8 +163,10 @@ parseCabalInfo cabalFileAsGiven sourceFileAsGiven = liftIO $ do
         }
     )
   where
-    whenNothing :: (Monad m) => Maybe a -> m a -> m a
+    whenNothing :: (Applicative f) => Maybe a -> f a -> f a
     whenNothing maya ma = maybe ma pure maya
+    whenLeft :: (Applicative f) => Either e a -> (e -> f a) -> f a
+    whenLeft eitha ma = either ma pure eitha
 
 -- | Get a map from Haskell source file paths (without any extensions) to
 -- the corresponding 'DynOption's and dependencies.